+++ projects/swimlane/ngx-ui/src/lib/components/input/input.component.ts
@@ -83,6 +83,7 @@ export class InputComponent implements AfterViewInit, OnDestroy, ControlValueAcc
   @Input() size: Size = Size.Small;
   @Input() appearance: Appearance = Appearance.Legacy;
   @Input() withMargin = true;
+  @Input() textareaRow = 1;
 
   @Input()
   get disabled() {