+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Enhancement: Added ARIA role attribute to `ngx-plus-menu`
 - Fix: Toggle going out of bounds when disabled in `ngx-toggle`
 - Fix: Use input password type for password fields in JSON editor.
+- Fix: Call to `onTouchedCallback()` should be executed on blur, per `ControlValueAccessor.registerOnTouched` docs
 - Fix: Tabs in `ngx-tabs` are now `type="button"`
 
 ## 35.6.8 (2021-07-16)