+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -11,6 +11,8 @@
 - Fix: Use input password type for password fields in JSON editor.
 - Fix: Call to `onTouchedCallback()` should be executed on blur, per `ControlValueAccessor.registerOnTouched` docs
 - Fix: Tabs in `ngx-tabs` are now `type="button"`
+- Fix: Adjust label position and background of `ngx-input` fill appearance
+- Enhancement: Add input, `textareaRow` to allow setting `row` attribute of `ngx-input` textarea
 
 ## 35.6.8 (2021-07-16)