+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -12,7 +12,7 @@
 - Fix: Call to `onTouchedCallback()` should be executed on blur, per `ControlValueAccessor.registerOnTouched` docs
 - Fix: Tabs in `ngx-tabs` are now `type="button"`
 - Fix: Adjust label position and background of `ngx-input` fill appearance
-- Enhancement: Add input, `textareaRow` to allow setting `row` attribute of `ngx-input` textarea
+- Enhancement: Add input, `textareaRows` to allow setting `row` attribute of `ngx-input` textarea
 
 ## 35.6.8 (2021-07-16)