+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## HEAD (unreleased)
 
+- Enhancement(Dropdown): Add viewport intersection to support opening select dropdowns if hidden by viewport.
 - Feature: Added focus rings to buttons
 - Feature: Add `passwordToggleEnabled` input to JSON editor.
 - Enhancement: Improve semantic HTML in `ngx-input` and `ngx-select`